Refactor the monstrosity
